refactor: moved render.js from components folder to scripts folder
